pyproject.toml
[tool.poetry]
name = "wrtn-data-pipeline"
version = "0.1.0"
description = ""
authors = ["sejeong-park <tpwjd29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.47.0"
mysql-connector-python = "^9.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
